add rollux breakdown by token to current chainge adapter
